sonic-agent/requirements.txt
# Core dependencies
kafka-python==2.0.2
pydantic==2.0.0
python-dotenv==1.0.0
tenacity==8.2.3
structlog==23.1.0

# Optional for monitoring
psutil==5.9.6
prometheus-client==0.18.0

# SONiC dependencies (already on system)
# sonic-py-common
# sonic-platform-common
